Changes Made
These changes however made 2 behviors:

1. On the initial open it sync's the contacts 2 times.
2. When launched FROM A LOCKED DEVICE it give the ready beep 2 times instead of once. Sometimes the 2 beeps are so close together it's difficult to hear that it is 2 beeps.

These are minor issues but if you know why that is happening and can fix it without breaking the ability to launch this successfully from a locked device.

Also, I detected a problem with the Contacts Sync on all the devices. With Initial launch it successfully syncs contacts. But the button to re-sync contacts pop's up a "Uploading Contacts" message but does not actually update the contacts that have changed.

After you have made the above changes I would like you to again post the code to GIT